Don't shadow exceptions
I'm not a huge fan of this cop:

https://www.rubydoc.info/gems/rubocop/RuboCop/Cop/Lint/ShadowedException

Mostly because I find the error extremely hard to interpret. But after
examining the errors on JWT, it does appear that VerificationError
inherits from DecodeError and thus specifying them both is redundant.

I left a comment because this wasn't clear to me and seemed worth
communicating to the next dev reading this code.