Extract helper function to get ids from order data
I found this spec hard to read because there was so much noise around
getting the ids from the order data. Some tests used strings to compare
the ids with the fabricated data and others converted the fabricated
data to integers. With this helper in place there's now one consistent
way to do this and thus the signal in these tests is easier to find.