Modify `rake graphql:schema:diff_check` to not depend on Git.
* Output some indication of the GraphQL schema check running

As CI runs within a Docker container, and we (sensibly) exclude the
.git/ directory from Docker builds, CI was wrongly failing when checking
for GraphQL schema diff as git was failing.

This commit modifies the diff check protocol to only depend on `cp` and
`diff`. Notably it leaves a _schema.graphql.orig file around in the
filesystem if there is diff that wasn't generated previously.