Clean up some RuboCop config
Rather than setting the max class length higher and higher, my thought
is to specify the files that are offenders in the todo file in the hopes
that we'll take another look at them to look for ways to make them
smaller. I have not disabled the ClassLength cop so that other files
will still cause RuboCop to complain, which is probably good for us!

/cc @ashkan18