Remove order edit button from active admin order page
Updating `actions` to `actions :all, except: %i[create destroy new]` in https://github.com/artsy/exchange/pull/636 added the `edit_admin_order_path` path (used for confirm sale button) as well as adding the actual button to page. That new `edit` button is not needed and also dangerous. This piece of code removes that. Source: https://stackoverflow.com/questions/44723359/how-do-you-remove-the-edit-button-in-an-active-admin-show